Removed I2S Tx, shrunk buffer sizes (unnecessarily)
Trying to get this to place and route.  Turns out that the 512x32 and the 256x32
dual port RAMS (with the other port being x128) AND the 1024x32 all take the
same number of BRAMs.  I'll be sizing them back up, I think.

Had to reduce the logic size, and removing I2S Tx seemed to do the trick.
Too bad, it would have been nice to have.